refactor: Modify the project structure
