+++ client/src/components/nav/MainNav.js
@@ -1,5 +1,5 @@
 import Nav from './Nav.js';
-import { convertStringToHTML } from '../../utils.js';
+import { convertStringToHTML } from '../../js/util';
 import { store } from '../../data/store.js';
 
 const MainNav = () => convertStringToHTML(`<nav class="contents main-nav">${Nav(store.navItems.main)}</nav>`);
