+++ client/src/components/section/DailySeriesRanking.js
@@ -1,6 +1,6 @@
 import { store } from '../../data/store.js';
 import { ContentsHeader, ContentsNav, ContentsList } from '../index.js';
-import { convertStringToHTML } from '../../utils.js';
+import { convertStringToHTML } from '../../js/util';
 
 const DailySeriesRanking = () => {
