+++ client/src/components/section/GenreTop.js
@@ -1,6 +1,6 @@
 import { store } from '../../data/store.js';
 import { ContentsHeader, ContentsList } from '../index.js';
-import { convertStringToHTML } from '../../utils.js';
+import { convertStringToHTML } from '../../js/util';
 
 const GenreTop = (genre) => {
