+++ client/src/js/event/index.js
@@ -1,0 +1,4 @@
+import { navCallback } from './navEvent.js';
+import { bannerCallback, setCarousel, carouselCallback, carouselTransitionHandler } from './bannerEvent.js';
+
+export { navCallback, bannerCallback, setCarousel, carouselCallback, carouselTransitionHandler };
