--- client/src/js/index.js
@@ -1,4 +1,0 @@
-import { navCallback } from './event/navEvent.js';
-import { bannerCallback, setCarousel, carouselCallback, carouselTransitionHandler } from './event/bannerEvent.js';
-
-export { navCallback, bannerCallback, setCarousel, carouselCallback, carouselTransitionHandler };
