+++ client/src/js/render/index.js
@@ -1,7 +1,7 @@
-import { $ } from './utils.js';
-import { Nav, PageNav, DailySeriesRanking, MainContents } from './components';
-import { PageComponent } from './data/components.js';
-import { store } from './data/store.js';
+import { $ } from '../util';
+import { Nav, PageNav, DailySeriesRanking, MainContents } from '../../components';
+import { PageComponent } from '../../data/components.js';
+import { store } from '../../data/store.js';
 
 const renderPageNav = () => {
   if ($('.page-nav') === null) $('.header-wrapper').insertAdjacentHTML('beforeend', PageNav());
